# Conflicts:
#	public_html/lists/admin/locale/templates/phplist.pot

+++ public_html/lists/admin/lib.php
@@ -114,6 +114,17 @@ function setMessageData($msgid, $name, $value)
             }
         }
     }
+    if ($name == 'excludelist' && is_array($value)) {
+        ## make sure all entries are numerical. @@TO(DO We could also check that they are actually valid list IDs
+        $newLists = array();
+        foreach ($value as $v) {
+            if (is_numeric($v)) {
+                $newLists[$v] = $v;
+            }
+        }
+        $value = $newLists;
+        unset($newLists);
+    }
     if (is_array($value) || is_object($value)) {
         $value = 'SER:'.serialize($value);
     }
@@ -969,22 +980,28 @@ function checkLock($processid)
 
 function getPageCache($url, $lastmodified = 0)
 {
+    if (empty($_SESSION['hasconf'])) return;
     $req = Sql_Fetch_Row_Query(sprintf('select content from %s where url = "%s" and lastmodified >= %d',
         $GLOBALS['tables']['urlcache'], $url, $lastmodified));
-
-    return $req[0];
+    if (!empty($req) && is_array($req)) {
+        return $req[0];
+    }
+    return '';
 }
 
 function getPageCacheLastModified($url)
 {
-    $req = Sql_Fetch_Row_Query(sprintf('select lastmodified from %s where url = "%s"', $GLOBALS['tables']['urlcache'],
-        $url));
-
-    return $req[0];
+    if (empty($_SESSION['hasconf'])) return;
+    $req = Sql_Fetch_Row_Query(sprintf('select lastmodified from %s where url = "%s"', $GLOBALS['tables']['urlcache'],$url));
+    if (!empty($req) && is_array($req)) {
+        return $req[0];
+    }
+    return 0;
 }
 
 function setPageCache($url, $lastmodified, $content)
 {
+    if (empty($_SESSION['hasconf'])) return;
     //  if (isset($GLOBALS['developer_email'])) return;
     Sql_Query(sprintf('delete from %s where url = "%s"', $GLOBALS['tables']['urlcache'], $url));
     Sql_Query(sprintf('insert into %s (url,lastmodified,added,content)
@@ -1133,7 +1150,7 @@ function testUrl($url)
     if (VERBOSE) {
         logEvent('Checking '.$url);
     }
-    $code = 500;
+
     if ($GLOBALS['has_curl']) {
         if (VERBOSE) {
             logEvent('Checking curl ');
@@ -1147,18 +1164,24 @@ function testUrl($url)
         curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, false);
         curl_setopt($curl, CURLOPT_HEADER, 0);
         curl_setopt($curl, CURLOPT_DNS_USE_GLOBAL_CACHE, true);
-        curl_setopt($curl, CURLOPT_USERAGENT, 'phplist v'.VERSION.' (https://www.phplist.com)');
+        curl_setopt($curl, CURLOPT_USERAGENT, 'phplist v'.VERSION.'c (https://www.phplist.com)');
         $raw_result = curl_exec($curl);
         $code = curl_getinfo($curl, CURLINFO_HTTP_CODE);
-    } elseif ($GLOBALS['has_pear_http_request']) {
+    } else {
         if (VERBOSE) {
-            logEvent('Checking PEAR ');
+            logEvent('Checking HTTP_Request2 ');
         }
-        @require_once 'HTTP/Request.php';
-        $headreq = new HTTP_Request($url, $request_parameters);
-        $headreq->addHeader('User-Agent', 'phplist v'.VERSION.' (https://www.phplist.com)');
-        if (!PEAR::isError($headreq->sendRequest(false))) {
-            $code = $headreq->getResponseCode();
+        require_once 'HTTP/Request2.php';
+
+        $headreq = new HTTP_Request2($url, HTTP_Request2::METHOD_HEAD, array('follow_redirects' => true));
+        $headreq->setHeader('User-Agent', 'phplist v'.VERSION.'p (https://www.phplist.com)');
+
+        try {
+            $response = $headreq->send();
+            $code = $response->getStatus();
+        } catch (HTTP_Request2_Exception $e) {
+            logEvent(sprintf('Error fetching %s %s', $url, $e->getMessage()));
+            $code = 500;
         }
     }
     if (VERBOSE) {
@@ -1168,7 +1191,19 @@ function testUrl($url)
     return $code;
 }
 
-function fetchUrl($url, $userdata = array())
+/**
+ * Returns the content of a URL from
+ *  the global variable $urlcache,
+ *  or the urlcache table
+ *  or fetching the URL directly.
+ *
+ * @param string $url      the URL to fetch
+ * @param array  $userdata user fields to be replaced in the url
+ * @param int    $ttl      time to live, the number of seconds after which the cached item will expire
+ *
+ * @return string|false the url content or false for an error
+ */
+function fetchUrl($url, $userdata = array(), $ttl = REMOTE_URL_REFETCH_TIMEOUT)
 {
     $content = '';
 
@@ -1193,7 +1228,7 @@ function fetchUrl($url, $userdata = array())
 
     // keep in memory cache in case we send a page to many emails
     if (isset($GLOBALS['urlcache'][$url]) && is_array($GLOBALS['urlcache'][$url])
-        && (time() - $GLOBALS['urlcache'][$url]['fetched'] < REMOTE_URL_REFETCH_TIMEOUT)
+        && (time() - $GLOBALS['urlcache'][$url]['fetched'] < $ttl)
     ) {
         //     logEvent($url . " is cached in memory");
         if (VERBOSE && function_exists('output')) {
@@ -1205,7 +1240,7 @@ function fetchUrl($url, $userdata = array())
 
     $dbcache_lastmodified = getPageCacheLastModified($url);
     $timeout = time() - $dbcache_lastmodified;
-    if ($timeout < REMOTE_URL_REFETCH_TIMEOUT) {
+    if ($timeout < $ttl) {
         //    logEvent($url.' was cached in database');
         if (VERBOSE && function_exists('output')) {
             output('From database cache: '.$url);
@@ -1228,29 +1263,21 @@ function fetchUrl($url, $userdata = array())
     //# see http://mantis.phplist.com/view.php?id=7684
 //    $lastmodified = strtotime($header["last-modified"]);
     $lastmodified = time();
-    $cache = getPageCache($url, $lastmodified);
-    if (!$cache) {
-        if (function_exists('curl_init')) {
-            $content = fetchUrlCurl($url, $request_parameters);
-        } elseif (0 && $GLOBALS['has_pear_http_request'] == 2) {
-            //# @#TODO, make it work with Request2
-            @require_once 'HTTP/Request2.php';
-        } elseif ($GLOBALS['has_pear_http_request']) {
-            @require_once 'HTTP/Request.php';
-            $content = fetchUrlPear($url, $request_parameters);
-        } else {
-            return false;
-        }
-    } else {
+    $content = getPageCache($url, $lastmodified);
+
+    if ($content) {
         if (VERBOSE) {
             logEvent($url.' was cached in database');
         }
-        $content = $cache;
+    } else {
+        $content = fetchUrlDirect($url, $request_parameters);
     }
 
     if (!empty($content)) {
         $content = addAbsoluteResources($content, $url);
-        logEvent('Fetching '.$url.' success');
+        if (VERBOSE) {
+            logEvent('Fetching '.$url.' success');
+        }
         setPageCache($url, $lastmodified, $content);
 
         $GLOBALS['urlcache'][$url] = array(
@@ -1260,6 +1287,31 @@ function fetchUrl($url, $userdata = array())
     }
 
     return $content;
+}
+
+/**
+ * Fetches a URL directly.
+ * Use curl if available otherwise fallback to HTTP_Request2.
+ *
+ * @param string  $url               the URL to fetch
+ * @param array   $requestParameters params for the http request
+ *
+ * @return string|false the url content or false for an error
+ */
+function fetchUrlDirect($url, $requestParameters = array())
+{
+    global $has_curl;
+
+    $defaultParameters = array(
+        'timeout' => 10,
+    );
+    $parameters = $requestParameters + $defaultParameters;
+
+    if ($has_curl) {
+        return fetchUrlCurl($url, $parameters);
+    }
+
+    return fetchUrlHttpRequest2($url, $parameters);
 }
 
 function fetchUrlCurl($url, $request_parameters)
@@ -1291,59 +1343,47 @@ function fetchUrlCurl($url, $request_parameters)
     }
 }
 
-function fetchUrlPear($url, $request_parameters)
+/**
+ * Fetches a URL using the PEAR package HTTP_Request2.
+ *
+ * @param string  $url               the URL to fetch
+ * @param array   $requestParameters params for the http request
+ *
+ * @return string the url content or false for an error
+ */
+function fetchUrlHttpRequest2($url, $requestParameters)
 {
+    require_once 'HTTP/Request2.php';
+
     if (VERBOSE) {
-        logEvent($url.' fetching with PEAR');
+        logEvent("Fetching $url with HTTP_Request2");
     }
+    $request = new HTTP_Request2(
+        $url,
+        HTTP_Request2::METHOD_GET,
+        array(
+            'timeout' => $requestParameters['timeout'],
+            'follow_redirects' => true,
+        )
+    );
+    $request->setHeader('User-Agent', 'phplist v'.VERSION.'p (http://www.phplist.com)');
 
-    if (0 && $GLOBALS['has_pear_http_request'] == 2) {
-        $headreq = new HTTP_Request2($url, $request_parameters);
-        $headreq->setHeader('User-Agent', 'phplist v'.VERSION.'p (https://www.phplist.com)');
-    } else {
-        $headreq = new HTTP_Request($url, $request_parameters);
-        $headreq->addHeader('User-Agent', 'phplist v'.VERSION.'p (https://www.phplist.com)');
-    }
-    if (!PEAR::isError($headreq->sendRequest(false))) {
-        $code = $headreq->getResponseCode();
-        if ($code != 200) {
-            logEvent('Fetching '.$url.' failed, error code '.$code);
+    try {
+        $response = $request->send();
 
-            return 0;
-        }
-        $header = $headreq->getResponseHeader();
+        if ($response->getStatus() == 200) {
+            $content = $response->getBody();
 
-        if (preg_match('/charset=(.*)/i', $header['content-type'], $regs)) {
-            $remote_charset = strtoupper($regs[1]);
-        }
-
-        $request_parameters['method'] = 'GET';
-        if (0 && $GLOBALS['has_pear_http_request'] == 2) {
-            $req = new HTTP_Request2($url, $request_parameters);
-            $req->setHeader('User-Agent', 'phplist v'.VERSION.'p (https://www.phplist.com)');
-        } else {
-            $req = new HTTP_Request($url, $request_parameters);
-            $req->addHeader('User-Agent', 'phplist v'.VERSION.'p (https://www.phplist.com)');
-        }
-        logEvent('Fetching '.$url);
-        if (VERBOSE && function_exists('output')) {
-            output('Fetching remote: '.$url);
-        }
-        if (!PEAR::isError($req->sendRequest(true))) {
-            $content = $req->getResponseBody();
-
-            if ($remote_charset != 'UTF-8' && function_exists('iconv')) {
-                $content = iconv($remote_charset, 'UTF-8//TRANSLIT', $content);
+            if (VERBOSE) {
+                logEvent("Fetched $url");
             }
         } else {
-            logEvent('Fetching '.$url.' failed on GET '.$req->getResponseCode());
-
-            return 0;
+            logEvent(sprintf('Unexpected HTTP status: %s %s', $response->getStatus(), $response->getReasonPhrase()));
+            $content = false;
         }
-    } else {
-        logEvent('Fetching '.$url.' failed on HEAD');
-
-        return 0;
+    } catch (HTTP_Request2_Exception $e) {
+        logEvent(sprintf('Error fetching %s %s', $url, $e->getMessage()));
+        $content = false;
     }
 
     return $content;
@@ -1809,10 +1849,10 @@ function refreshTlds($force = 0)
         //# even if it fails we mark it as done, so that we won't getting stuck in eternal updating.
         SaveConfig('tld_last_sync', time(), 0);
         if (defined('TLD_AUTH_LIST')) {
-            $tlds = fetchUrl(TLD_AUTH_LIST);
+            $tlds = fetchUrlDirect(TLD_AUTH_LIST);
         }
         if ($tlds && defined('TLD_AUTH_MD5')) {
-            $tld_md5 = fetchUrl(TLD_AUTH_MD5);
+            $tld_md5 = fetchUrlDirect(TLD_AUTH_MD5);
             list($remote_md5, $fname) = explode(' ', $tld_md5);
             $mymd5 = md5($tlds);
 //        print 'OK: '.$remote_md5.' '.$mymd5;
@@ -1868,7 +1908,7 @@ function shortenTextDisplay($text, $max = 30)
         return mb_shortenTextDisplay($text, $max);
     }
 
-    $text = str_replace('http://', '', $text);
+    $text = preg_replace('!^https?://!i', '', $text);
     if (strlen($text) > $max) {
         if ($max < 30) {
             $display = substr($text, 0, $max - 4).' ... ';
@@ -1881,13 +1921,13 @@ function shortenTextDisplay($text, $max = 30)
     $display = str_replace('/', '/&#x200b;', $display);
     $display = str_replace('@', '@&#x200b;', $display);
 
-    return sprintf('<span title="%s" ondblclick="alert(\'%s\');">%s</span>', htmlspecialchars($text),
+    return sprintf('<span title="%s">%s</span>', htmlspecialchars($text),
         htmlspecialchars($text), $display);
 }
 
 function mb_shortenTextDisplay($text, $max = 30)
 {
-    $text = str_replace('http://', '', $text);
+    $text = preg_replace('!^https?://!i', '', $text);
     if (mb_strlen($text) > $max) {
         if ($max < 30) {
             $display = mb_substr($text, 0, $max - 4).' ... ';
@@ -2309,4 +2349,19 @@ function asyncLoadContentDiv($url,$divname)
         asyncLoadDiv[asyncLoadDiv.length] = "'.$divname.'";
         asyncLoadUrl[asyncLoadUrl.length] = "'.$url.'";
      </script>';
+}
+
+/**
+ * Transform a value to be valid for an html id by removing invalid characters.
+ * This is for HTML 4. HTML 5 is more lenient.
+ *
+ * @see https://www.w3.org/TR/html4/types.html#type-id
+ *
+ * @param string $value
+ *
+ * @return string
+ */
+function sanitiseId($value)
+{
+    return preg_replace('/[^0-9A-Za-z\-_:.]/', '', $value);
 }
